Move CBC tests into read strategies
